% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{caviarpd-package}
\alias{caviarpd-package}
\title{caviarpd: Cluster Analysis via Random Partition Distributions}
\description{
Cluster analysis is performed using pairwise distance information and a random partition distribution. The method is implemented for two random partition distributions. It draws samples and then obtains and plots clustering estimates. An implementation of a selection algorithm is provided for the mass parameter of the partition distribution. Since pairwise distances are the principal input to this procedure, it is most comparable to the hierarchical and k-medoids clustering methods. The method is Dahl, Andros, Carter (2022+) \doi{10.1002/sam.11602}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbdahl/caviarpd-package}
  \item Report bugs at \url{https://github.com/dbdahl/caviarpd-package/issues}
}

}
\author{
\strong{Maintainer}: David B. Dahl \email{dahl@stat.byu.edu} (\href{https://orcid.org/0000-0002-8173-1547}{ORCID})

Authors:
\itemize{
  \item R. Jacob Andros \email{androsrj@gmail.com} (\href{https://orcid.org/0000-0002-1289-385X}{ORCID})
  \item J. Brandon Carter \email{carterj4@icloud.com} (\href{https://orcid.org/0000-0003-1687-0564}{ORCID})
}

Other contributors:
\itemize{
  \item Authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
