% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDCRPPartition.R
\name{DDCRPPartition}
\alias{DDCRPPartition}
\title{Probabilities for the Distance-Dependent CRP (ddCRP) Partition Distribution}
\usage{
DDCRPPartition(similarity, concentration)
}
\arguments{
\item{similarity}{An n-by-n symmetric matrix giving the similarity between
pairs of items.  Matrix entries must be strictly positive.}

\item{concentration}{The concentration parameter, as a numeric value
greater than \eqn{-1*discount}.}
}
\value{
An object of class \code{PartitionDistribution} representing this
partition distribution.
}
\description{
This function specifies the Distance-Dependent CRP (ddCRP) partition
distribution for given similarity matrix and concentration.
}
\examples{
subset <- c("Maine","Georgia","California","Minnesota","Montana")
d <- as.matrix(dist(scale(USArrests[subset,])))
temperature <- 1.0
similarity <- exp( -temperature * d )
DDCRPPartition(similarity=similarity, concentration=0.4)
}
