% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpmm.R
\name{dpmm_fit}
\alias{dpmm_fit}
\title{Sample from the Posterior Partition Distribution of the Neal (2000) Model}
\usage{
dpmm_fit(
  data,
  n_samples,
  burnin,
  thin,
  mass = 1,
  sd = 0.1,
  algorithm = "neal8"
)
}
\value{
A matrix in which each column represents the partition of the items.
}
\description{
This function samples from the posterior partition distribution of the Neal (2000) Model.  See Section 8 of
"Markov Chain Sampling Methods for Dirichlet Process Mixture Models" by Radford M. Neal,
Journal of Computational and Graphical Statistics, Vol. 9, No. 2. (Jun., 2000), pp. 249-265.
}
\examples{
x <- dpmm_generate(10)
dpmm_fit(x$data, 100, 10, 2)

}
