% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpmm.R
\name{dpmm_generate}
\alias{dpmm_generate}
\title{Generate Data from the Neal (2000) Model}
\usage{
dpmm_generate(n_items, mass = 1, sd = 0.1)
}
\value{
A list giving the true partition, the associated cluster parameters, and the simulated data.
}
\description{
This function generates data from the model in Section 8 of "Markov Chain Sampling Methods for Dirichlet Process Mixture Models" by Radford M. Neal,
Journal of Computational and Graphical Statistics, Vol. 9, No. 2. (Jun., 2000), pp. 249-265.
}
\examples{
dpmm_generate(10)

}
