% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPAPartition.R
\name{EPAPartition}
\alias{EPAPartition}
\title{Probabilities for the Ewens-Pitman Attaction (EPA) Partition Distribution}
\usage{
EPAPartition(similarity, permutation, concentration, discount = 0)
}
\arguments{
\item{similarity}{An n-by-n symmetric matrix giving the similarity between
pairs of items.  Matrix entries must be strictly positive.}

\item{permutation}{A vector containing the integers \eqn{1, 2, \ldots, n}
giving the order in which items are allocated to the partition.}

\item{concentration}{The concentration parameter, as a numeric value
greater than \eqn{-1*discount}.}

\item{discount}{The discount parameter, as a numeric value in \eqn{[0,1)}.}
}
\value{
An object of class \code{PartitionDistribution} representing this
partition distribution.
}
\description{
This function specifies the Ewens-Pitman attraction partition distribution
given similarity matrix, permutation, concentration, and
discount parameters.
}
\examples{
subset <- c("Maine","Georgia","California","Minnesota","Montana")
d <- as.matrix(dist(scale(USArrests[subset,])))
temperature <- 1.0
similarity <- exp( -temperature * d )
EPAPartition(similarity=similarity, permutation=c(1,5,4,2,3), concentration=1.5, discount=0.1)
}
