% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.part.f.R
\name{est.part.f}
\alias{est.part.f}
\title{Fernando's Splintered Clustering Idea}
\usage{
est.part.f(post.mean)
}
\arguments{
\item{post.mean}{Vector containing estimated posterior means for those subjects under consideration.}
}
\value{
A list giving the number of clusters, cluster labels, and cluster sizes using \code{\link[mclust:Mclust]{mclust::Mclust()}}.
}
\description{
Fernando's Splintered Clustering Idea
}
