% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{gourd-package}
\alias{gourd}
\alias{gourd-package}
\title{gourd: Shrinkage Partition Distribution}
\description{
Implementation of the distribution introduced in our paper "Dependent Random Partitions by Shrinking Toward an Anchor".
}
\author{
\strong{Maintainer}: David B. Dahl \email{dahl@stat.byu.edu} (\href{https://orcid.org/0000-0002-8173-1547}{ORCID})

Authors:
\itemize{
  \item Richard L. Warr \email{warr@stat.byu.edu} (\href{https://orcid.org/0000-0001-8508-3105}{ORCID})
  \item Thomas P. Jensen \email{tompjensen@gmail.com} (\href{https://orcid.org/0009-0000-0015-3881}{ORCID})
}

Other contributors:
\itemize{
  \item Authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
