% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsh_loss_builder_process.R
\name{gsh_loss_builder_process}
\alias{gsh_loss_builder_process}
\title{Add Bayesian Networks to GSH Builder Structure}
\usage{
gsh_loss_builder_process(x, gsh_loss_builder)
}
\arguments{
\item{x}{A network as an adjacency matrix.}

\item{gsh_loss_builder}{A pointer to a GSH builder structure.}
}
\description{
This function adds a Bayesian network represented by an adjacency matrix to an existing GSH
builder structure.
}
\examples{
data(choi25)
gsh_builder <- gsh_loss_builder_new(ncol(choi25), nrow(choi25), 4, a = 1)
gsh_loss_builder_process(choi25[,,1], gsh_builder)

}
\references{
D. B. Dahl, E. Bailey, D. Harris, J. Andros (2025), Bayesian Network Estimation from Posterior Samples,
\url{https://dahl.byu.edu/papers/dahl_bailey_harris_andros.pdf}.
}
