% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dag.R
\name{is_dag}
\alias{is_dag}
\title{Check if a Graph is a Directed Acyclic Graph (DAG)}
\usage{
is_dag(candidate)
}
\arguments{
\item{candidate}{An adjacency matrix representing a candidate graph.}
}
\value{
A logical value indicating whether or not the graph is a DAG.
}
\description{
This function checks whether the provided graph is a Directed Acyclic Graph (DAG).
}
\examples{
# Example 1: This demonstration checks if a real DAG returns true.
dag_matrix <- matrix(c(0, 1, 0,
                       0, 0, 1,
                       0, 0, 0), nrow = 3, byrow = TRUE)
is_dag(dag_matrix)

# Example 2: This demonstration checks if a graph with a cycle returns false.
cycle_matrix <- matrix(c(0, 1, 0,
                         0, 0, 1,
                         1, 0, 0), nrow = 3, byrow = TRUE)
is_dag(cycle_matrix)

}
