% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{lorenzpartition-package}
\alias{lorenzpartition}
\alias{lorenzpartition-package}
\title{lorenzpartition: Lorenz Partition Distribution}
\description{
Implements the Lorenz partition distribution, a prior on set partitions that separates the number of clusters k, the integer partition (cluster sizes) given k, and the allocation of items to clusters. The integer partition component uses a sequential kernel construction (TiDaL or TaDPoLe) whose target is specified via Lorenz curve increments. Functions are provided to evaluate the probability mass function ('dlorenzip'), draw random samples ('rlorenzip'), construct Lorenz curves ('lorenz_linear', 'lorenz_ispline'), and run diagnostic comparisons ('downsample', 'crp_vs_lorenz'). See Dahl, Warr, and Hilton (2025+) for details.
}
\author{
\strong{Maintainer}: David B. Dahl \email{dahl@stat.byu.edu} (\href{https://orcid.org/0000-0002-8173-1547}{ORCID})

Authors:
\itemize{
  \item Richard L. Warr \email{warr@stat.byu.edu} (\href{https://orcid.org/0000-0001-8508-3105}{ORCID})
  \item Talmage Hilton \email{hilto24@student.byu.edu}
}

Other contributors:
\itemize{
  \item Authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
