% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nealAlgorithm.R
\name{nealAlgorithm8}
\alias{nealAlgorithm8}
\title{Neal's Algorithm 8 for a Partition}
\usage{
nealAlgorithm8(
  partition,
  parameters,
  distr,
  logLikelihood = function(i, parameter) 0,
  sampleParameter = function() 0,
  updateParameter = function(parameter, subset) 0,
  mcmcTuning = list()
)
}
\arguments{
\item{partition}{An integer vector of cluster labels representing the current
partition. Items \eqn{i} and \eqn{j} are in the same cluster if and only if
\code{partition[i] == partition[j]}.}

\item{parameters}{A list whose \eqn{j}th element is the shared model
parameter for all those items having cluster label \eqn{j} in the
\code{partition} argument.}

\item{distr}{A specification of the partition distribution, i.e., an object
of class \code{PartitionDistribution} as returned by, for example, a
function such as \code{\link{CRPPartition}}.}

\item{logLikelihood}{A function taking an index \eqn{i} (as a single
integer), a cluster \eqn{label} (as a single integer), and an
indication as to whether a new parameter should be sampled (as an single
logical) and returning the natural logarithm of \eqn{p( y_i | parameter
  associated with "label" )}, i.e., that item's contribution to the log
likelihood given the parameter associated with the cluster \dQuote{label}.}

\item{sampleParameter}{A function taking no arguments and returning a draw
from the prior parameter distribution (i.e., the centering distribution).}

\item{updateParameter}{A function taking a \code{parameter} (a current value
of a parameter) and a subset of integers \code{subset} (as a numeric
vector) and returning an updated value for the parameter by some valid MCMC
update scheme for the full conditional distribution of the parameter.}

\item{mcmcTuning}{An optional list containing \code{nUpdates} (an integer
giving the number of Gibbs scans before returning, defaulting to \eqn{1})
and other variables specific to the particular prior partition
distribution.}
}
\value{
A list containing \code{partition} (an integer vector giving the
updated partition encoded using cluster labels), \code{parameters} (a list
like the \code{parameters} argument, with updated values), and possibly
other items specific to a particular partition distribution.
}
\description{
This function performs Algorithm 8 of Neal (2000), which updates a partition
based on a partition prior distribution and user-supplied functions to
evaluate an item's contribution to the loglikelihood sample a model parameter
for its prior distribution, and update a model parameter based on a subset.
}
\examples{
# Save current seed
if ( exists(".Random.seed") ) {
  savedSeed <- .Random.seed
  rm(.Random.seed, envir=globalenv())
}

# Neal (2000) model and data
data <- c(-1.48, -1.40, -1.16, -1.08, -1.02, 0.14, 0.51, 0.53, 0.78)
sigma2 <- 0.1^2
mu0 <- 0
sigma02 <- 1

## For Neal Algorithm 3
logIntegratedLikelihoodItem <- function(i, subset) {
  posteriorVariance     <- 1 / ( 1/sigma02 + length(subset)/sigma2 )
  posteriorMean         <- posteriorVariance * ( mu0/sigma02 + sum(data[subset])/sigma2 )
  posteriorPredictiveSD <- sqrt(posteriorVariance + sigma2)
  dnorm(data[i], posteriorMean, posteriorPredictiveSD, log=TRUE)
}

## For Neal Algorithm 8
logLikelihoodOfItem <- function(i, parameter) {
  dnorm(data[i], mean=parameter, sd=sqrt(sigma2), log=TRUE)
}

sampleParameter <- function() {
  rnorm(1, mean=mu0, sd=sqrt(sigma02))
}

updateParameter <- function(parameter, subset) {
  posteriorVariance <- 1 / ( 1/sigma02 + length(subset)/sigma2 )
  posteriorMean     <- posteriorVariance * ( mu0/sigma02 + sum(data[subset])/sigma2 )
  rnorm(1, mean=posteriorMean, sd=sqrt(posteriorVariance))
}

# Prior partition distribution.  Many are supported.
whichDistr <- sample(c("up","jlp","crp","epa","lsp","cpp","sp"),1)
cat("Using selected distribution: ", whichDistr, "\n", sep="")
concentration <- 1.5
discount <- 0.2
anchor <- c(1,1,2,2,2,3,4,4,4)
permutation <- sample(length(anchor))
distr <- if ( whichDistr == "up" ) {
  UniformPartition(nItems=length(data))
} else if ( whichDistr == "jlp" ) {
  JensenLiuPartition(concentration=concentration, permutation=permutation)
} else if ( whichDistr == "crp" ) {
  CRPPartition(nItems=length(data), concentration=concentration, discount=discount)
} else if ( whichDistr == "epa" ) {
  subset <- c("Maine","Georgia","California","Minnesota","Oregon",
              "Utah","Wisconsin","Texas","Nevada")
  d <- as.matrix(dist(scale(USArrests[subset,])))
  temperature <- 1.0
  similarity <- exp( -temperature * d )
  EPAPartition(similarity, permutation, concentration, discount)
} else if ( whichDistr == "lsp" ) {
  randomizeShrinkage <- sample(c("fixed","common"),1)
  shrinkage <- 2
  LocationScalePartition(anchor, shrinkage=shrinkage,
                         concentration=concentration, permutation=permutation)
} else if ( whichDistr == "cpp" ) {
  randomizeShrinkage <- "fixed"
  shrinkage <- 2
  CenteredPartition(anchor=anchor, shrinkage=shrinkage,
                    baseline=CRPPartition(nItems=length(data),
                                          concentration=concentration, discount=discount))
} else if ( whichDistr == "sp" ) {
  randomizeShrinkage <- sample(c("fixed","common","cluster","idiosyncratic"),1)
  shrinkage <- if ( randomizeShrinkage == "common" ) c(0.2,0.2,0,0,0,0.2,0.2,0.2,0.2)
  else if ( randomizeShrinkage == "cluster" ) c(0.2,0.2,0,0,0,0.1,0.4,0.4,0.4)
  else if ( randomizeShrinkage == "idiosyncratic" ) c(0.2,0.2,0,0,0,0.1,0.4,0.4,0.4)
  else if ( randomizeShrinkage == "fixed" ) c(0.2,0.2,0,0,0,0.1,0.4,0.4,0.4)
  else stop("Unrecognized 'randomizeShrinkage'")
  ShrinkagePartition(anchor, shrinkage=shrinkage, permutation, grit=0.2,
                     baseline=CRPPartition(nItems=length(data),
                                           concentration=concentration, discount=discount))
} else stop("Unrecongized distribution")

# MCMC

getOr <- function(x, or) if ( is.null(x) ) or else x
mcmcTuning <- list(nUpdates=2)
nSamples <- 100L
partitionSamples <- matrix(1, nrow=nSamples, ncol=length(data))
parametersSamples <- list(list(sampleParameter()))
whichIndex <- 1
shrinkageSamples <- matrix(getOr(distr$shrinkage,0), nrow=nSamples, ncol=length(data))
concentrationSamples <- rep(getOr(if ( inherits(distr,"ShrinkagePartition") ) {
    distr$baseline$concentration
  } else {
    distr$concentration
  }, 0), nSamples)
discountSamples <- rep(getOr(if ( inherits(distr,"ShrinkagePartition") ) {
    distr$baseline$discount
  } else {
    distr$discount
  }, 0), nSamples)
nAcceptancesPermutation <- nAcceptancesConcentration <- nAcceptancesDiscount <- 0L

doAlgorithm3 <- sample(c(TRUE,FALSE),1)
for ( i in 2:nSamples ) {
  # Update partition
  if ( doAlgorithm3 ) {
    partitionSamples[i,] <- nealAlgorithm3(partitionSamples[i-1,], distr,
                                           logIntegratedLikelihoodItem, mcmcTuning)$partition
  } else {
    out <- nealAlgorithm8(partitionSamples[i-1,], parametersSamples[[i-1]], distr,
                          logLikelihoodOfItem, sampleParameter, updateParameter, mcmcTuning)
    partitionSamples[i,] <- out$partition
    parametersSamples[[i]] <- out$parameters
  }
  # Update permutation
  out <- updatePermutation(distr, partitionSamples[i,], k=length(data)/2)
  distr <- out$distr
  nAcceptancesPermutation <- nAcceptancesPermutation + out$accepted
  # Update shrinkage
  distr <- updateShrinkage(distr, partitionSamples[i,], randomizeShrinkage=randomizeShrinkage)
  shrinkageSamples[i,] <- getOr(distr$shrinkage,0)
  # Update concentration
  out <- updateConcentration(distr, partitionSamples[i,], rwsd=2)
  distr <- out$distr
  nAcceptancesConcentration <- nAcceptancesConcentration + out$accepted
  concentrationSamples[i] <- if ( inherits(distr,"ShrinkagePartition") ) {
    getOr(distr$baseline$concentration,0)
  } else {
    getOr(distr$concentration,0)
  }
  # Update discount
  out <- updateDiscount(distr, partitionSamples[i,], rwsd=0.1)
  distr <- out$distr
  nAcceptancesDiscount <- nAcceptancesDiscount + out$accepted
  discountSamples[i] <- if ( inherits(distr,"ShrinkagePartition") ) {
    getOr(distr$baseline$discount,0)
  } else {
    getOr(distr$discount,0)
  }
  # Perhaps add code to update other parameters
}

# Acceptance rate for updates
nAcceptancesPermutation / (nSamples-1)
nAcceptancesConcentration / (nSamples-1)

nSubsets <- apply(partitionSamples, 1, function(x) length(unique(x)))
plot(nSubsets, type="l")
mean(nSubsets)
sum(acf(nSubsets)$acf)-1   # Autocorrelation time

plot(shrinkageSamples[,1], type="l", main = "Trace Plot of Shrinkage")
plot(shrinkageSamples[,3], type="l", main = "Trace Plot of Shrinkage")
plot(shrinkageSamples[,6], type="l", main = "Trace Plot of Shrinkage")
plot(shrinkageSamples[,7], type="l", main = "Trace Plot of Shrinkage")
plot(shrinkageSamples[,8], type="l", main = "Trace Plot of Shrinkage")

plot(concentrationSamples, type="l", main = "Trace Plot of Concentration")

# Restore seed
if ( exists("savedSeed") ) .Random.seed <- savedSeed

}
\seealso{
\code{\link{CRPPartition}}, \code{\link{ShrinkagePartition}},
\code{\link{LocationScalePartition}}, \code{\link{CenteredPartition}},
\code{\link{prPartition}}, \code{\link{samplePartition}},
\code{\link{updatePermutation}}, \code{\link{nealAlgorithm3}}
}
