% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan}
\alias{plan}
\title{Plan Splinters for the Splintered Clustering Procedure}
\usage{
plan(n_items, n_splinters, splinter_size, n_anchors = 0)
}
\arguments{
\item{n_items}{Total number of items to cluster.}

\item{n_splinters}{Number of splinters to form.}

\item{splinter_size}{Number of items in each splinter.}

\item{n_anchors}{Number of anchors, i.e., items found in each splinter.}
}
\value{
A list containing \sQuote{nSplinters} lists, each containing a vector
\sQuote{indices} giving the item numbers for a given splinter.
}
\description{
This function provides a randomize splintering plan for the splintered
clustering procedure. Note that \sQuote{nSplinters} * \sQuote{splinterSize}
must be at least \sQuote{nItems}.
}
\examples{
plan(20, 5, 8)

}
