% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic_spline.R
\name{plot.cubic_spline}
\alias{plot.cubic_spline}
\title{Plot a Cubic Spline}
\usage{
\method{plot}{cubic_spline}(x, n_points = 201L, ...)
}
\arguments{
\item{x}{A cubic spline object (functions as the \code{spline} parameter).}

\item{n_points}{Number of points to use for plotting (default 201).}

\item{...}{Additional arguments passed to \code{\link[=plot]{plot()}}.}
}
\value{
Invisibly returns the cubic spline object.
}
\description{
Plots a cubic spline curve.
}
\examples{
# Create and plot a curved spline
spline <- cubic_spline(c(0, 0.2, 0.5, 0.8, 1), c(5, 10, 25, 40, 50))
plot(spline, main = "Concentration Curve")

}
