% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{pumpkin-package}
\alias{pumpkin}
\alias{pumpkin-package}
\title{pumpkin: Random Partition Models}
\description{
An implementation of probability mass functions and sampling algorithms for various random partition models are provided.
}
\author{
\strong{Maintainer}: David B. Dahl \email{dahl@stat.byu.edu} (\href{https://orcid.org/0000-0002-8173-1547}{ORCID})

Other contributors:
\itemize{
  \item Authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
