% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{slicesampler-package}
\alias{slicesampler}
\alias{slicesampler-package}
\title{slicesampler: Slice Sampler}
\description{
Neal's sampler sampler (Neal 2003) is implemented..
}
\author{
\strong{Maintainer}: David B. Dahl \email{dahl@stat.byu.edu} (\href{https://orcid.org/0000-0002-8173-1547}{ORCID})

Other contributors:
\itemize{
  \item Authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
