% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_loss.R
\name{splinclust_expected_loss}
\alias{splinclust_expected_loss}
\title{Computed Expected Loss from Splinters}
\usage{
splinclust_expected_loss(estimate, splinters, useVI = TRUE, a = 1)
}
\arguments{
\item{estimate}{A numeric vector giving an estimated clustering.}

\item{splinters}{A nonempty list containing lists.  Each list must be of
length two containing, as its first element, a numeric vector named
\dQuote{indices} and, as its second element, a numeric matrix named
\dQuote{partitions} with number of rows equaling the length of
\dQuote{indices} and at least one column.}

\item{useVI}{Use the (generalized) variation of information loss?  If not,
Binder loss is used.}

\item{a}{The argument \code{a} is a nonnegative scalar in \eqn{[0,2]} giving
(for Binder loss) the cost of placing two items in separate clusters when in
truth they belong to the same cluster. Without loss of generality, the cost
(under Binder loss) of placing two items in the same cluster when in truth
they belong to separate clusters is fixed \code{2-a}. For VI, \code{a} has a
similar interpretation, although is not a unit cost. See Dahl, Johnson,
Müller (2022).}
}
\value{
A numeric vector.
}
\description{
This function computes the expected loss from splinters.
}
\examples{
cat("To do.\n")

}
