% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{treeSB-package}
\alias{treeSB}
\alias{treeSB-package}
\title{treeSB: A Short Phrase in Title Case}
\description{
One paragraph description of what the package does.
}
\author{
\strong{Maintainer}: First Last \email{name@gmail.com} (\href{https://orcid.org/ORCID-ID}{ORCID})

Other contributors:
\itemize{
  \item Authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
