% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_slice_sampler_doubling.R
\name{univariate_slice_sampler_doubling}
\alias{univariate_slice_sampler_doubling}
\title{Univariate Slice Sampler using Doubling Method and Shrinkage}
\usage{
univariate_slice_sampler_doubling(x, target, on_log_scale, width)
}
\arguments{
\item{x}{A numeric scalar giving the current value.}

\item{target}{A function taking a numeric scalar and returning the target function evaluated at that value.}

\item{on_log_scale}{A logical scalar indicating whether \code{target} is on the log scale.}

\item{width}{A numeric scalar giving the estimated slice width.}
}
\value{
A numeric vector.
}
\description{
This function implements a slice sampler using the doubling and shrinkage
methods.
}
\examples{
n_samples <- 1000
sum <- 0.0
state <- 0.5
target <- function(x) if (0 < x && x < 1) x else 0
for (i in seq_len(n_samples)) {
  state <- univariate_slice_sampler_doubling(state, target, FALSE, 1)$value
  sum <- sum + state
}
sum / n_samples

}
\references{
Radford M Neal. (2003). Slice Sampling.
\emph{The Annals of Statistics}. Vol 31, No 3, 705-767.
}
