% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nealAlgorithmCorrelatedPartitions.R
\name{updateAnchorOfCorrelatedPartitions}
\alias{updateAnchorOfCorrelatedPartitions}
\title{Update the Anchor Partition of Correlated Partitions}
\usage{
updateAnchorOfCorrelatedPartitions(distr, correlated)
}
\arguments{
\item{distr}{A specification of the partition distribution, i.e., an object
of class \code{PartitionDistribution} as returned by, for example, a
function such as \code{\link{CRPPartition}}.}

\item{correlated}{A list of lists, with each list containing two elements: i.
a partition (in cluster label notation), and ii. a Shrinkage partition
distribution.}
}
\description{
This function updates the common anchor partitions in a collection of
shrinkage partition distributions.
}
