% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDiscount.R
\name{updateDiscount}
\alias{updateDiscount}
\title{Update the Discount Parameter of a Partition Distribution}
\usage{
updateDiscount(
  distr,
  partition,
  rwsd = 0.1,
  logPriorDensity = function(w) dunif(w, 0, 1, log = TRUE)
)
}
\arguments{
\item{distr}{A specification of the partition distribution, i.e., an object
of class \code{PartitionDistribution} as returned by, for example, a
function such as \code{\link{CRPPartition}}.}

\item{partition}{An integer vector of cluster labels representing the current
partition. Items \eqn{i} and \eqn{j} are in the same cluster if and only if
\code{partition[i] == partition[j]}.}

\item{rwsd}{Standard deviation of the Gaussian random walk proposal density
for the Metropolis algorithm.}

\item{logPriorDensity}{A function taking a numeric scalar and returning the
log of the prior density evaluated at the value.}
}
\value{
A list containing the potentially updated partition distribution and
a logical named \code{accepted} (indicating whether the proposal was
accepted).
}
\description{
This function updates the discount parameter of a partition distribution (e.g,
\code{\link{CRPPartition}} and \code{\link{ShrinkagePartition}}) using a
Gaussian random walk.  The prior can be specified by the user.
}
\examples{
# Save current seed
if ( exists(".Random.seed") ) {
  savedSeed <- .Random.seed
  rm(.Random.seed, envir=globalenv())
}

# Neal (2000) model and data
data <- c(-1.48, -1.40, -1.16, -1.08, -1.02, 0.14, 0.51, 0.53, 0.78)
sigma2 <- 0.1^2
mu0 <- 0
sigma02 <- 1

## For Neal Algorithm 3
logIntegratedLikelihoodItem <- function(i, subset) {
  posteriorVariance     <- 1 / ( 1/sigma02 + length(subset)/sigma2 )
  posteriorMean         <- posteriorVariance * ( mu0/sigma02 + sum(data[subset])/sigma2 )
  posteriorPredictiveSD <- sqrt(posteriorVariance + sigma2)
  dnorm(data[i], posteriorMean, posteriorPredictiveSD, log=TRUE)
}

## For Neal Algorithm 8
logLikelihoodOfItem <- function(i, parameter) {
  dnorm(data[i], mean=parameter, sd=sqrt(sigma2), log=TRUE)
}

sampleParameter <- function() {
  rnorm(1, mean=mu0, sd=sqrt(sigma02))
}

updateParameter <- function(parameter, subset) {
  posteriorVariance <- 1 / ( 1/sigma02 + length(subset)/sigma2 )
  posteriorMean     <- posteriorVariance * ( mu0/sigma02 + sum(data[subset])/sigma2 )
  rnorm(1, mean=posteriorMean, sd=sqrt(posteriorVariance))
}

# Prior partition distribution.  Many are supported.
whichDistr <- sample(c("up","jlp","crp","epa","lsp","cpp","sp"),1)
cat("Using selected distribution: ", whichDistr, "\n", sep="")
concentration <- 1.5
discount <- 0.2
anchor <- c(1,1,2,2,2,3,4,4,4)
permutation <- sample(length(anchor))
distr <- if ( whichDistr == "up" ) {
  UniformPartition(nItems=length(data))
} else if ( whichDistr == "jlp" ) {
  JensenLiuPartition(concentration=concentration, permutation=permutation)
} else if ( whichDistr == "crp" ) {
  CRPPartition(nItems=length(data), concentration=concentration, discount=discount)
} else if ( whichDistr == "epa" ) {
  subset <- c("Maine","Georgia","California","Minnesota","Oregon",
              "Utah","Wisconsin","Texas","Nevada")
  d <- as.matrix(dist(scale(USArrests[subset,])))
  temperature <- 1.0
  similarity <- exp( -temperature * d )
  EPAPartition(similarity, permutation, concentration, discount)
} else if ( whichDistr == "lsp" ) {
  randomizeShrinkage <- sample(c("fixed","common"),1)
  shrinkage <- 2
  LocationScalePartition(anchor, shrinkage=shrinkage,
                         concentration=concentration, permutation=permutation)
} else if ( whichDistr == "cpp" ) {
  randomizeShrinkage <- "fixed"
  shrinkage <- 2
  CenteredPartition(anchor=anchor, shrinkage=shrinkage,
                    baseline=CRPPartition(nItems=length(data),
                                          concentration=concentration, discount=discount))
} else if ( whichDistr == "sp" ) {
  randomizeShrinkage <- sample(c("fixed","common","cluster","idiosyncratic"),1)
  shrinkage <- if ( randomizeShrinkage == "common" ) c(0.2,0.2,0,0,0,0.2,0.2,0.2,0.2)
  else if ( randomizeShrinkage == "cluster" ) c(0.2,0.2,0,0,0,0.1,0.4,0.4,0.4)
  else if ( randomizeShrinkage == "idiosyncratic" ) c(0.2,0.2,0,0,0,0.1,0.4,0.4,0.4)
  else if ( randomizeShrinkage == "fixed" ) c(0.2,0.2,0,0,0,0.1,0.4,0.4,0.4)
  else stop("Unrecognized 'randomizeShrinkage'")
  ShrinkagePartition(anchor, shrinkage=shrinkage, permutation, grit=0.2,
                     baseline=CRPPartition(nItems=length(data),
                                           concentration=concentration, discount=discount))
} else stop("Unrecongized distribution")

# MCMC

getOr <- function(x, or) if ( is.null(x) ) or else x
mcmcTuning <- list(nUpdates=2)
nSamples <- 100L
partitionSamples <- matrix(1, nrow=nSamples, ncol=length(data))
parametersSamples <- list(list(sampleParameter()))
whichIndex <- 1
shrinkageSamples <- matrix(getOr(distr$shrinkage,0), nrow=nSamples, ncol=length(data))
concentrationSamples <- rep(getOr(if ( inherits(distr,"ShrinkagePartition") ) {
    distr$baseline$concentration
  } else {
    distr$concentration
  }, 0), nSamples)
discountSamples <- rep(getOr(if ( inherits(distr,"ShrinkagePartition") ) {
    distr$baseline$discount
  } else {
    distr$discount
  }, 0), nSamples)
nAcceptancesPermutation <- nAcceptancesConcentration <- nAcceptancesDiscount <- 0L

doAlgorithm3 <- sample(c(TRUE,FALSE),1)
for ( i in 2:nSamples ) {
  # Update partition
  if ( doAlgorithm3 ) {
    partitionSamples[i,] <- nealAlgorithm3(partitionSamples[i-1,], distr,
                                           logIntegratedLikelihoodItem, mcmcTuning)$partition
  } else {
    out <- nealAlgorithm8(partitionSamples[i-1,], parametersSamples[[i-1]], distr,
                          logLikelihoodOfItem, sampleParameter, updateParameter, mcmcTuning)
    partitionSamples[i,] <- out$partition
    parametersSamples[[i]] <- out$parameters
  }
  # Update permutation
  out <- updatePermutation(distr, partitionSamples[i,], k=length(data)/2)
  distr <- out$distr
  nAcceptancesPermutation <- nAcceptancesPermutation + out$accepted
  # Update shrinkage
  distr <- updateShrinkage(distr, partitionSamples[i,], randomizeShrinkage=randomizeShrinkage)
  shrinkageSamples[i,] <- getOr(distr$shrinkage,0)
  # Update concentration
  out <- updateConcentration(distr, partitionSamples[i,], rwsd=2)
  distr <- out$distr
  nAcceptancesConcentration <- nAcceptancesConcentration + out$accepted
  concentrationSamples[i] <- if ( inherits(distr,"ShrinkagePartition") ) {
    getOr(distr$baseline$concentration,0)
  } else {
    getOr(distr$concentration,0)
  }
  # Update discount
  out <- updateDiscount(distr, partitionSamples[i,], rwsd=0.1)
  distr <- out$distr
  nAcceptancesDiscount <- nAcceptancesDiscount + out$accepted
  discountSamples[i] <- if ( inherits(distr,"ShrinkagePartition") ) {
    getOr(distr$baseline$discount,0)
  } else {
    getOr(distr$discount,0)
  }
  # Perhaps add code to update other parameters
}

# Acceptance rate for updates
nAcceptancesPermutation / (nSamples-1)
nAcceptancesConcentration / (nSamples-1)

nSubsets <- apply(partitionSamples, 1, function(x) length(unique(x)))
plot(nSubsets, type="l")
mean(nSubsets)
sum(acf(nSubsets)$acf)-1   # Autocorrelation time

plot(shrinkageSamples[,1], type="l", main = "Trace Plot of Shrinkage")
plot(shrinkageSamples[,3], type="l", main = "Trace Plot of Shrinkage")
plot(shrinkageSamples[,6], type="l", main = "Trace Plot of Shrinkage")
plot(shrinkageSamples[,7], type="l", main = "Trace Plot of Shrinkage")
plot(shrinkageSamples[,8], type="l", main = "Trace Plot of Shrinkage")

plot(concentrationSamples, type="l", main = "Trace Plot of Concentration")

# Restore seed
if ( exists("savedSeed") ) .Random.seed <- savedSeed

}
\seealso{
\code{\link{nealAlgorithm3}}, \code{\link{nealAlgorithm8}}
}
